Gen tweaks + Fixed ore detector + Silo Terminal return + Cliff Explosive

+++ src/fos/type/blocks/production/OreDetector.java
@@ -76,15 +76,13 @@ public class OreDetector extends Block {
                 Draw.z(Layer.bullet - 0.0001f);
                 Draw.alpha(0.6f);
                 Lines.stroke(2.5f, Color.valueOf("4b95ff"));
+                Draw.alpha(0.2f);
                 float x2 = x + (Mathf.cos(Time.time / 18f) * range());
                 float y2 = y + (Mathf.sin(Time.time / 18f) * range());
-                Draw.z(Layer.block - 1f);
                 Lines.line(x, y, x2, y2);
-                Draw.z(Layer.bullet - 0.0001f);
-                Draw.alpha(0.2f);
                 Drawf.circles(x, y, range(), Color.valueOf("4b95ff"));
                 Drawf.circles(x, y, range() * 0.95f, Color.valueOf("4b95ff"));
-                locateOres((int)x, (int)y, range());
+                locateOres(range());
             }
         }
 
@@ -93,20 +91,20 @@ public class OreDetector extends Block {
             Drawf.dashCircle(x, y, range, Color.valueOf("4b95ff"));
         }
 
-        public void locateOres(int x, int y, float range) {
-            for (float i = -range; i <= range * 4; i+=8) {
-                for (float j = -range; j <= range * 4; j+=8) {
-                    Tile tile = world.tileWorld(i, j);
+        public void locateOres(float range) {
+            for (float i = -range; i <= range; i+=8) {
+                for (float j = -range; j <= range; j+=8) {
+                    Tile tile = world.tileWorld(x + i, y + j);
                     //oh god so many conditions here
-                    if (Mathf.within(x, y, i, j, range) && tile != null && tile.overlay() != null && tile.overlay() instanceof UndergroundOreBlock) {
+                    if (Mathf.within(x, y, x + i, y + j, range) && tile != null && tile.overlay() != null && tile.overlay() instanceof UndergroundOreBlock) {
                         Draw.z(1f);
                         Draw.alpha(0.4f);
                         Drawf.square(tile.x * 8, tile.y * 8, 3, Mathf.PI / 4, tile.drop().color);
 
-                        //show an item icon above the cursor/finger?
+                        //show an item icon above the cursor/finger
                         Tile hoverTile = world.tileWorld(Core.input.mouseWorld().x, Core.input.mouseWorld().y);
 
-                        if (tile == hoverTile) {
+                        if (tile == hoverTile && tile.block() != null) {
                             Draw.z(Layer.max);
                             Draw.alpha(1f);
                             Draw.rect(tile.drop().uiIcon, tile.x * 8, tile.y * 8 + 8);
